libinjection: Splited the code into separeted header files.
