+++ libraries/libinjection/Container.h
@@ -2,175 +2,17 @@
 
 #include <libtest/Assets.h>
 
-#include <libutils/Callback.h>
 #include <libutils/HashMap.h>
-#include <libutils/RefPtr.h>
-#include <libutils/String.h>
-#include <libutils/Traits.h>
 #include <libutils/TypeId.h>
 #include <libutils/Vector.h>
 
+#include <libinjection/Entity.h>
+#include <libinjection/Factories.h>
+#include <libinjection/Lifetimes.h>
+#include <libinjection/Retriever.h>
+
 namespace Injection
 {
-
-struct Container;
-
-struct Entity : public AnyRef
-{
-};
-
-struct Context
-{
-private:
-    Vector<TypeId> _stack;
-    Container &_container;
-
-public:
-    Container &container() const { return _container; }
-
-    Context(Container &container)
-        : _container{container}
-    {
-    }
-
-    template <typename T>
-    void push_type()
-    {
-        _stack.push_back(GetTypeId<T>());
-    }
-
-    void ensure_no_cycle()
-    {
-        for (size_t i = 0; i < _stack.count() - 1; ++i)
-        {
-            assert_not_equal(_stack[i], _stack.peek_back());
-        }
-    }
-
-    template <typename T>
-    void pop_type()
-    {
-        assert_equal(_stack.pop_back(), GetTypeId<T>());
-    }
-};
-
-template <typename T>
-struct ConstantFactory
-{
-private:
-    RefPtr<T> _instance;
-
-public:
-    ConstantFactory(RefPtr<T> instance) : _instance(instance)
-    {
-    }
-
-    RefPtr<T> instance(Context &)
-    {
-        return _instance;
-    }
-};
-
-template <typename T>
-struct ConstructorFactory
-{
-    RefPtr<T> instance(Context &context)
-    {
-        return make<T>(context);
-    }
-};
-
-template <typename T>
-struct CallbackFactory
-{
-private:
-    Callback<T()> _callback;
-
-public:
-    RefPtr<T> instance(Context &context)
-    {
-        return _callback(context);
-    }
-};
-
-template <typename TInstance, typename TFactory>
-struct TransientLifeTime : public RefCounted<TransientLifeTime<TInstance, TFactory>>
-{
-private:
-    TFactory _factorie;
-
-public:
-    TransientLifeTime(TFactory factorie)
-        : _factorie{factorie}
-    {
-    }
-
-    virtual RefPtr<TInstance> instance(Context &context)
-    {
-        context.push_type<TInstance>();
-
-        context.ensure_no_cycle();
-
-        auto instance = _factorie.instance(context);
-
-        context.pop_type<TInstance>();
-
-        return instance;
-    }
-};
-
-template <typename TInstance, typename TFactory>
-struct SingletonLifeTime : public RefCounted<SingletonLifeTime<TInstance, TFactory>>
-{
-private:
-    TFactory _factorie;
-    RefPtr<TInstance> _instance;
-
-public:
-    SingletonLifeTime(TFactory factorie)
-        : _factorie{factorie}
-    {
-    }
-
-    virtual RefPtr<TInstance> instance(Context &context)
-    {
-        context.push_type<TInstance>();
-
-        context.ensure_no_cycle();
-
-        if (_instance == nullptr)
-        {
-            _instance = _factorie.instance(context);
-        }
-
-        context.pop_type<TInstance>();
-
-        return _instance;
-    }
-};
-
-template <typename TInterface>
-struct Retriever : public AnyRef
-{
-    virtual RefPtr<TInterface> instance(Context &context) = 0;
-};
-
-template <typename TInterface, typename TInstance>
-struct CastRetriever : public Retriever<TInterface>
-{
-private:
-    RefPtr<TInstance> _storage;
-
-public:
-    CastRetriever(RefPtr<TInstance> storage) : _storage(storage)
-    {
-    }
-
-    RefPtr<TInterface> instance(Context &context) override
-    {
-        return _storage->instance(context);
-    }
-};
 
 struct Container
 {
@@ -178,10 +20,6 @@ private:
     HashMap<TypeId, Vector<RefPtr<AnyRef>>> _retrievers;
 
 public:
-    Container() {}
-
-    ~Container() {}
-
     template <typename TLifetime, typename... TInterfaces>
     requires(sizeof...(TInterfaces) == 0) void register_interfaces(RefPtr<TLifetime>) {}
 
@@ -229,9 +67,9 @@ public:
     }
 
     template <typename TInterface>
-    RefPtr<TInterface> get()
+    requires(!IsVector<TInterface>::value) RefPtr<TInterface> get()
     {
-        auto id = GetTypeId<TInterface>();
+        TypeId id = GetTypeId<TInterface>();
 
         if (!_retrievers.has_key(id))
         {
@@ -245,13 +83,15 @@ public:
     }
 
     template <typename TInterface>
-    void get(Vector<RefPtr<TInterface>> &instances)
+    Vector<RefPtr<TInterface>> get_all()
     {
-        auto id = GetTypeId<TInterface>();
+        Vector<RefPtr<TInterface>> instances;
+
+        TypeId id = GetTypeId<TInterface>();
 
         if (!_retrievers.has_key(id))
         {
-            return;
+            return instances;
         }
 
         Vector<RefPtr<AnyRef>> &retrievers = _retrievers[id];
@@ -261,6 +101,8 @@ public:
             Context context{*this};
             instances.push_back(static_cast<RefPtr<Retriever<TInterface>>>(retrievers[i])->instance(context));
         }
+
+        return instances;
     }
 };
