+++ libraries/libinjection/Entity.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <libutils/RefCounted.h>
+
+namespace Injection
+{
+
+struct Entity : public AnyRef
+{
+};
+
+} // namespace Injection
